requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
backports-zoneinfo==0.2.1
    # via django
build==0.10.0
    # via pip-tools
click==8.1.3
    # via pip-tools
django==4.2.1
    # via
    #   -r requirements.in
    #   djangorestframework
django-environ==0.10.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
mysqlclient==1.4.6
    # via -r requirements.in
packaging==23.1
    # via build
pip-tools==6.13.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via build
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
